Only the owner can CRUD the habits/journals
Only the owner of a habit, measured habit, or journal entry can see,
edit, delete, or create the given object for that user. This change adds
tests to make sure we are covering this now and in the future (of
course, new code will not be covered by these checks).